backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
openai==1.3.7
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-allauth
python-decouple
PyPDF2
python-docx